// fe_v2/src/lib/api/generated/model/userSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface UserSummary {
  capabilities: string[];
  designations: string[];
  email: string;
  first_name: string;
  id: string;
  last_name: string;
  role: string;
  student_lifecycle_status: string | null;
}